使用库进行调试
